requirements.txt
beautifulsoup4
requests
fp-free-proxy
GitPython